requirements/requirements-base.txt
wheel

django
djangorestframework
django-filter

oauth2client
requests

# For Elasticsearch 6.0 and later, use the major version 6 (6.x.y) of the library.
elasticsearch>=6.0.0,<7.0.0

google-api-python-client
